requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
gunicorn==20.1.0